refactor: move changeCowName to its own file
